changed the search handling and all that stuff again.
It works now!

However I dont know yet how to avoid exceptions when asking for a field that doesnt exist in one of my JSON Objects. Sometimes there isnt a rating for a place. Maybe if statements work fine? I gotta test that.

For now PlaceInfo only contains the Name and the ID